move LocalEnvManager to ml-agents-envs
